Strong and weak validators are one module that keeps them apart
`ETag`, `Last-Modified`, and the `If-None-Match` comparison, with the
distinction that matters kept visible in the names.
`extract_validators_from_response` answers what a response offered;
`extract_strong_validators_from_response` answers the narrower question of what
a `Range` or `If-Range` may turn on. A weak validator is a validator and is
still not one of those, so the two are separate functions rather than one with
a flag.

`normalize_etag` comes along and its warning comes with it: stripping `W/` *is*
weak comparison, and the prefix it discards is exactly what a strong comparison
turns on — so it must never be used to prepare one.

Removed on the way out: a `#[cfg(test)] mod validator_extraction_tests` left
holding nothing but four `use` lines once its tests moved.

headers.rs: 2344 → 788 non-test lines, from 70 top-level items to 20.
