+++ lint-http-rules/src/helpers/mod.rs
@@ -62,5 +62,6 @@ pub mod status;
 pub mod structured_fields;
 pub mod token;
 pub mod uri;
+pub mod validator;
 pub mod vary;
 pub mod websocket;
