+++ lint-http-rules/src/rules/if_none_match_etag_syntax.rs
@@ -104,7 +104,7 @@ impl Rule for IfNoneMatchEtagSyntax {
                         "If-None-Match header contains an empty list element".into(),
                     ));
                 }
-                if let Err(msg) = crate::helpers::headers::validate_entity_tag(member) {
+                if let Err(msg) = crate::helpers::validator::validate_entity_tag(member) {
                     return Some(self.violation(
                         ctx.severity,
                         format!(
